requirements.txt
streamlit>=1.28.0
pymupdf>=1.23.0
pdf2image>=1.16.0
pillow>=10.0.0
pytesseract>=0.3.10
langchain>=0.1.0
langchain-core>=0.1.0
groq>=0.9.0
langgraph>=0.2.0
python-dotenv>=1.0.0
